refactor: make user service less visible
